Add files via upload
Student Management System  Project
INTRODUCTION
Student Management System is software which is helpful for students as well as theschool authorities. In the current system all the activities are done manually. It is very timeconsuming and costly. Our Student Management System deals with the various activitiesrelated to the students.
There are mainly 2 modules in this software
1.User module
2.Student Module
In the Software we can register as a user and user has of two types, student andadministrator. Administrator has the power to add new user and can edit and delete a user. 
A student can register as user and can add edit and delete his profile.